changed write_wave to byte buffer

+++ cubesatsim/cubesatsim.h
@@ -263,7 +263,7 @@ int ctr = 0;
 int rd = 0;
 int nrd;
 void write_to_buffer(int i, int symbol, int val);
-void write_wave(int i, short int * buffer);
+void write_wave(int i, byte * buffer);
 void sleep(float time);
 int uart_fd;
